Removed vec implementation.
